+++ cli/tests/test_git_clone.rs
@@ -65,7 +65,6 @@ fn test_git_clone() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: uuqppmxq 3711b3b5 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -173,7 +172,6 @@ fn test_git_clone() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/nested/path/to/repo"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: vzqnnsmr fea36bca (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -273,7 +271,6 @@ fn test_git_clone_colocate() -> TestResult {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: uuqppmxq 3711b3b5 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -418,7 +415,6 @@ fn test_git_clone_colocate() -> TestResult {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/nested/path/to/repo"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: vzqnnsmr fea36bca (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -443,7 +439,6 @@ fn test_git_clone_colocate_via_config() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -470,7 +465,6 @@ fn test_git_clone_no_colocate() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -611,7 +605,6 @@ fn test_git_clone_tags() {
     bookmark: main@origin [new] tracked
     tag: v1.0@git [new] 
     tag: v2.0@git [new] 
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 88542a00 (empty) (no description set)
     Parent commit (@-)      : lnmyztun e93ca54d main | message
     Added 2 files, modified 0 files, removed 0 files
@@ -623,7 +616,6 @@ fn test_git_clone_tags() {
     Fetching into new repo in "$TEST_ENV/included"
     bookmark: main@origin [new] tracked
     tag: v2.0@git [new] 
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: uuqppmxq 676b2fd8 (empty) (no description set)
     Parent commit (@-)      : lnmyztun e93ca54d main | message
     Added 2 files, modified 0 files, removed 0 files
@@ -636,7 +628,6 @@ fn test_git_clone_tags() {
     bookmark: main@origin [new] tracked
     tag: v1.0@git [new] 
     tag: v2.0@git [new] 
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: pmmvwywv cd5996a2 (empty) (no description set)
     Parent commit (@-)      : lnmyztun e93ca54d main | message
     Added 2 files, modified 0 files, removed 0 files
@@ -647,7 +638,6 @@ fn test_git_clone_tags() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/none"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: rzvqmyuk 61c45a3c (empty) (no description set)
     Parent commit (@-)      : lnmyztun e93ca54d main | message
     Added 2 files, modified 0 files, removed 0 files
@@ -686,7 +676,6 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     Fetching into new repo in "$TEST_ENV/clone1"
     bookmark: feature1@origin [new] tracked
     bookmark: main@origin     [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 feature1 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -701,13 +690,6 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     [EOF]
     ");
 
-    // "trunk()" alias should be set to default bookmark "main"
-    let output = clone_dir1.run_jj(["config", "list", "--repo", "revset-aliases.'trunk()'"]);
-    insta::assert_snapshot!(output, @r#"
-    revset-aliases.'trunk()' = "main@origin"
-    [EOF]
-    "#);
-
     // Only the default bookmark will be imported if auto-track-bookmarks = '~*'
     test_env.add_config("remotes.origin.auto-track-bookmarks = '~*'");
     let output = root_dir.run_jj(["git", "clone", "source", "clone2"]);
@@ -716,8 +698,7 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     Fetching into new repo in "$TEST_ENV/clone2"
     bookmark: feature1@origin [new] untracked
     bookmark: main@origin     [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
-    Working copy  (@) now at: rzvqmyuk 27e56779 (empty) (no description set)
+    Working copy  (@) now at: pmmvwywv ea9c2659 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 feature1@origin main | message
     Added 1 files, modified 0 files, removed 0 files
     [EOF]
@@ -738,8 +719,7 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     Fetching into new repo in "$TEST_ENV/clone3"
     bookmark: feature1@origin [new] tracked
     bookmark: main@origin     [new] untracked
-    Setting the revset alias `trunk()` to `feature1@origin`
-    Working copy  (@) now at: nppvrztz b16020e9 (empty) (no description set)
+    Working copy  (@) now at: zxsnswpr 56139679 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 feature1 main@origin | message
     Added 1 files, modified 0 files, removed 0 files
     [EOF]
@@ -751,13 +731,6 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     main@origin: qomsplrm ebeb70d8 message
     [EOF]
     ");
-
-    // "trunk()" alias should be set to new default bookmark "feature1"
-    let output = clone_dir3.run_jj(["config", "list", "--repo", "revset-aliases.'trunk()'"]);
-    insta::assert_snapshot!(output, @r#"
-    revset-aliases.'trunk()' = "feature1@origin"
-    [EOF]
-    "#);
 
     // No bookmarks should be imported if both auto-track-bookmarks and
     // track-default-bookmark-on-clone are turned off
@@ -773,8 +746,7 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     Fetching into new repo in "$TEST_ENV/clone4"
     bookmark: feature1@origin [new] untracked
     bookmark: main@origin     [new] untracked
-    Setting the revset alias `trunk()` to `feature1@origin`
-    Working copy  (@) now at: wmwvqwsz 5068d576 (empty) (no description set)
+    Working copy  (@) now at: spxsnpux 1f499908 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 feature1@origin main@origin | message
     Added 1 files, modified 0 files, removed 0 files
     [EOF]
@@ -804,8 +776,7 @@ fn test_git_clone_remote_default_bookmark() -> TestResult {
     bookmark: feature1@origin [new] tracked
     bookmark: main@origin     [new] tracked
     Hint: `git.track-default-bookmark-on-clone=false` has no effect if `git.auto-local-bookmark` is enabled.
-    Setting the revset alias `trunk()` to `feature1@origin`
-    Working copy  (@) now at: vzqnnsmr fea36bca (empty) (no description set)
+    Working copy  (@) now at: wmwvqwsz 5068d576 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 feature1 main | message
     Added 1 files, modified 0 files, removed 0 files
     [EOF]
@@ -847,18 +818,9 @@ fn test_git_clone_remote_default_bookmark_with_escape() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: "\""@origin [new] tracked
-    Setting the revset alias `trunk()` to `"\""@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 "\"" | message
     Added 1 files, modified 0 files, removed 0 files
-    [EOF]
-    "#);
-
-    // "trunk()" alias should be escaped and quoted
-    let clone_dir = test_env.work_dir("clone");
-    let output = clone_dir.run_jj(["config", "list", "--repo", "revset-aliases.'trunk()'"]);
-    insta::assert_snapshot!(output, @r#"
-    revset-aliases.'trunk()' = '"\""@origin'
     [EOF]
     "#);
 }
@@ -877,7 +839,6 @@ fn test_git_clone_ignore_working_copy() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     [EOF]
     "#);
     let clone_dir = test_env.work_dir("clone");
@@ -934,7 +895,6 @@ fn test_git_clone_with_remote_name() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@upstream [new] tracked
-    Setting the revset alias `trunk()` to `main@upstream`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -988,7 +948,6 @@ fn test_git_clone_trunk_deleted() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1000,11 +959,6 @@ fn test_git_clone_trunk_deleted() {
     ------- stderr -------
     Forgot 1 local bookmarks.
     Forgot 1 remote bookmarks.
-    Warning: Failed to check mutability of the new working-copy revision.
-    Caused by:
-    1: Invalid `revset-aliases.immutable_heads()`
-    2: Revision `main@origin` doesn't exist
-    Hint: Use `jj config edit --repo` to adjust the `trunk()` alias.
     [EOF]
     ");
 
@@ -1015,11 +969,6 @@ fn test_git_clone_trunk_deleted() {
     ○  qomsplrm someone@example.org 1970-01-01 11:00:00 ebeb70d8
     │  message
     ◆  zzzzzzzz root() 00000000
-    [EOF]
-    ------- stderr -------
-    Warning: Failed to resolve `revset-aliases.trunk()`: Revision `main@origin` doesn't exist
-    The `trunk()` alias is temporarily set to `root()`.
-    Hint: Use `jj config edit --repo` to adjust the `trunk()` alias.
     [EOF]
     ");
 }
@@ -1081,7 +1030,6 @@ fn test_git_clone_conditional_config() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/new"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: zxsnswpr 5479cd52 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1124,7 +1072,6 @@ fn test_git_clone_with_depth() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1169,7 +1116,6 @@ fn test_git_clone_invalid_immutable_heads() {
     Caused by:
     1: Invalid `revset-aliases.immutable_heads()`
     2: Revision `unknown` doesn't exist
-    Setting the revset alias `trunk()` to `main@origin`
     Warning: Failed to check mutability of the new working-copy revision.
     Caused by:
     1: Invalid `revset-aliases.immutable_heads()`
@@ -1197,7 +1143,6 @@ fn test_git_clone_malformed() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Internal error: Failed to check out commit 2f4286212884d472a0b2013a961b695a144ac65c
     Caused by: Reserved path component .jj in $TEST_ENV/clone/.jj
     [EOF]
@@ -1208,7 +1153,7 @@ fn test_git_clone_malformed() {
     let output = clone_dir.run_jj(["status"]);
     insta::assert_snapshot!(output, @"
     ------- stderr -------
-    Error: The working copy is stale (not updated since operation 95b4c4e3a6ed).
+    Error: The working copy is stale (not updated since operation d43f9414f97f).
     Hint: Run `jj workspace update-stale` to update it.
     See https://docs.jj-vcs.dev/latest/working-copy/#stale-working-copy for more information.
     [EOF]
@@ -1258,7 +1203,6 @@ fn test_git_clone_with_global_git_remote_config() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1335,7 +1279,6 @@ fn test_git_clone_branch_or_tag() {
     ------- stderr -------
     Fetching into new repo in "$TEST_ENV/clone"
     bookmark: main@origin [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: sqpuoqvx 1ca44815 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1348,7 +1291,6 @@ fn test_git_clone_branch_or_tag() {
     insta::assert_snapshot!(repo_dir.run_jj(["config", "list", "--repo"]), @r#"
     remotes.origin.fetch-bookmarks = "main"
     remotes.origin.fetch-tags = "~*"
-    revset-aliases."trunk()" = "main@origin"
     [EOF]
     "#);
 
@@ -1436,7 +1378,6 @@ fn test_git_clone_branch_or_tag() {
     Fetching into new repo in "$TEST_ENV/clone_all"
     bookmark: feature1@origin [new] tracked
     bookmark: main@origin     [new] tracked
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: wmwvqwsz 5068d576 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1458,7 +1399,6 @@ fn test_git_clone_branch_or_tag() {
     bookmark: main@origin [new] tracked
     tag: tag2@origin [new] 
     tag: tag3@origin [new] 
-    Setting the revset alias `trunk()` to `main@origin`
     Working copy  (@) now at: uuzqqzqu c871b515 (empty) (no description set)
     Parent commit (@-)      : qomsplrm ebeb70d8 main | message
     Added 1 files, modified 0 files, removed 0 files
@@ -1469,7 +1409,6 @@ fn test_git_clone_branch_or_tag() {
     insta::assert_snapshot!(repo_dir.run_jj(["config", "list", "--repo"]), @r#"
     remotes.origin.fetch-bookmarks = "~feature1"
     remotes.origin.fetch-tags = "~tag1"
-    revset-aliases."trunk()" = "main@origin"
     [EOF]
     "#);